+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Native scan timer support for the Arduino Uno R4 (Renesas RA4M1) family.
+- Fixed several compiler warnings
+
 ## [2.0.3]
 ### Fixed
 - The row scan timing was too long on ESP32 platforms. Adjusted this timing to remove visible blinking.