+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.1.0]
 ### Added
 - Native scan timer support for the Arduino Uno R4 (Renesas RA4M1) family.
 - Fixed several compiler warnings
@@ -83,7 +84,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## 1.0.0 - 2017-12-24
 Initial release
 
-[Unreleased]: https://github.com/michaelkamprath/ShiftRegisterLEDMatrixLib/compare/v2.0.3...HEAD
+[Unreleased]: https://github.com/michaelkamprath/ShiftRegisterLEDMatrixLib/compare/v2.1.0...HEAD
+[2.1.0]: https://github.com/michaelkamprath/ShiftRegisterLEDMatrixLib/compare/v2.0.3...v2.1.0
 [2.0.3]: https://github.com/michaelkamprath/ShiftRegisterLEDMatrixLib/compare/v2.0.2...v2.0.3
 [2.0.2]: https://github.com/michaelkamprath/ShiftRegisterLEDMatrixLib/compare/v2.0.1...v2.0.2
 [2.0.1]: https://github.com/michaelkamprath/ShiftRegisterLEDMatrixLib/compare/v2.0.0...v2.0.1